backend/requirements.txt
fastapi
uvicorn
pydantic
python-jose
passlib[bcrypt]
fastapi-mail
python-dotenv
motor
jose
tensorflow
numpy
pillow
python-multipart
passlib
pymongo